refactor: de-monolithize renpho-api.ts into cohesive modules
