+++ src/services/renpho-api/constants.ts
@@ -1,0 +1,4 @@
+export const API_BASE = 'https://cloud.renpho.com';
+export const ENCRYPTION_SECRET = 'ed*wijdi$h6fe3ew';
+export const DEFAULT_PAGE_SIZE = 200;
+export const MAX_MEASUREMENT_SCAN = 1000;
